server/chain/x/decogit/types/message_create_sticker_test.go
package types

import (
	"testing"

	"decogit/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCreateSticker_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateSticker
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateSticker{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateSticker{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
